move ColorButton to components
