+++ app/shared/ui-settings/src/commonMain/kotlin/ui/settings/components/ColorButton.kt
@@ -1,0 +1,101 @@
+/*
+ * Copyright (C) 2024-2025 OpenAni and contributors.
+ *
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
+ *
+ * https://github.com/open-ani/ani/blob/main/LICENSE
+ */
+
+package me.him188.ani.app.ui.settings.components
+
+import androidx.compose.animation.core.animateDpAsState
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.sizeIn
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Check
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.drawBehind
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import com.materialkolor.hct.Hct
+import com.materialkolor.ktx.toHct
+
+
+@Composable
+fun ColorButton(
+    onClick: () -> Unit,
+    baseColor: Color,
+    selected: Boolean,
+    modifier: Modifier = Modifier,
+    cardColor: Color = MaterialTheme.colorScheme.surfaceContainer,
+    containerColor: Color = MaterialTheme.colorScheme.primaryContainer,
+) {
+    val containerSize by animateDpAsState(targetValue = if (selected) 28.dp else 0.dp)
+    val iconSize by animateDpAsState(targetValue = if (selected) 16.dp else 0.dp)
+
+    Surface(
+        modifier = modifier
+            .padding(4.dp)
+            .sizeIn(maxHeight = 80.dp, maxWidth = 80.dp, minHeight = 64.dp, minWidth = 64.dp)
+            .aspectRatio(1f),
+        shape = RoundedCornerShape(16.dp),
+        color = cardColor,
+        onClick = onClick,
+    ) {
+        Box(Modifier.fillMaxSize()) {
+            val hct = baseColor.toHct()
+            val color1 = Color(Hct.from(hct.hue, 40.0, 80.0).toInt())
+            val color2 = Color(Hct.from(hct.hue, 40.0, 90.0).toInt())
+            val color3 = Color(Hct.from(hct.hue, 40.0, 60.0).toInt())
+
+            Box(
+                modifier = modifier
+                    .size(48.dp)
+                    .clip(CircleShape)
+                    .drawBehind { drawCircle(color1) }
+                    .align(Alignment.Center),
+            ) {
+                Surface(
+                    color = color2,
+                    modifier = Modifier
+                        .align(Alignment.BottomStart)
+                        .size(24.dp),
+                ) {}
+                Surface(
+                    color = color3,
+                    modifier = Modifier
+                        .align(Alignment.BottomEnd)
+                        .size(24.dp),
+                ) {}
+                Box(
+                    modifier = Modifier
+                        .align(Alignment.Center)
+                        .clip(CircleShape)
+                        .size(containerSize)
+                        .drawBehind { drawCircle(containerColor) },
+                ) {
+                    Icon(
+                        imageVector = Icons.Outlined.Check,
+                        contentDescription = null,
+                        modifier = Modifier.size(iconSize).align(Alignment.Center),
+                        tint = MaterialTheme.colorScheme.onPrimaryContainer,
+                    )
+                }
+            }
+        }
+    }
+}
